seminar/exercises/datamining/core/src/test/java/de/tuberlin/dima/aim/exercises/one/PrimeNumbersWritableTest.java
/**
 * Copyright (C) 2011 AIM III course DIMA TU Berlin
 *
 * This programm is free software; you can redistribute it and/or modify
 * it under the terms of the Apache License, Version 2.0
 * (the "License"); you may not use this file except in compliance with
 * the License.  You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package de.tuberlin.dima.aim.exercises.one;

import org.junit.Test;

import java.io.ByteArrayInputStream;
import java.io.ByteArrayOutputStream;
import java.io.DataInputStream;
import java.io.DataOutputStream;
import java.io.IOException;

import static org.junit.Assert.assertEquals;

public class PrimeNumbersWritableTest {

  @Test
  public void writeAndRead() throws IOException {

    PrimeNumbersWritable original = new PrimeNumbersWritable(2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 53,
        59, 61, 67, 71, 73, 79, 83, 89, 97);

    ByteArrayOutputStream out = new ByteArrayOutputStream();
    original.write(new DataOutputStream(out));

    PrimeNumbersWritable clone = new PrimeNumbersWritable();
    clone.readFields(new DataInputStream(new ByteArrayInputStream(out.toByteArray())));

    assertEquals(original, clone);
  }

}
